[IPS] Address comments from review
