+++ ips_client/tools/anonymize_file.py
@@ -4,7 +4,6 @@ from pathlib import Path
 from requests.exceptions import ConnectionError
 from typing import Optional, Union
 
-from ips_client.data_models import Region
 from ips_client.job import IPSJob, JobArguments, ServiceType, OutputType
 from ips_client.settings import Settings
 from ips_client.tools.utils import normalize_path
@@ -17,10 +16,9 @@ settings = Settings()
 log.debug(f'Settings: {settings}')
 
 
-def anonymize_file(file_path: str, out_type: OutputType, service: ServiceType, region: Region = Region.european_union,
-                   face: bool = True, license_plate: bool = True, ips_url: str = settings.ips_url_default,
-                   out_path: Optional[str] = None, skip_existing: bool = True, save_metadata: bool = True,
-                   auto_delete_job: bool = True):
+def anonymize_file(file_path: str, out_type: OutputType, service: ServiceType, job_args: JobArguments = JobArguments(),
+                   ips_url: str = settings.ips_url_default, out_path: Optional[str] = None, skip_existing: bool = True,
+                   save_metadata: bool = True, auto_delete_job: bool = True):
     """
     If no out_path is given, <input_filename_anonymized> will be used.
     """
@@ -35,11 +33,8 @@ def anonymize_file(file_path: str, out_type: OutputType, service: ServiceType, r
         log.info(f'Skipping because output already exists: {out_path}')
         return
 
-    # assemble job arguments
-    job_args = JobArguments(region=region, face=face, license_plate=license_plate)
-    log.info(f'Job arguments: {job_args}')
-
     # anonymize
+    log.info(f'Job arguments: {job_args}')
     try:
         with open(file_path, 'rb') as file:
             job: IPSJob = IPSJob.start_new(file=file,
@@ -66,12 +61,11 @@ def anonymize_file(file_path: str, out_type: OutputType, service: ServiceType, r
 
 
 def _get_out_path(out_path: Union[str, Path], file_path: str) -> str:
-    if not out_path:
-        file_path = Path(file_path)
-        return normalize_path(
-            Path(file_path.parent).joinpath(f'{file_path.stem}_anonymized{file_path.suffix}')
-        )
-    return normalize_path(out_path)
+    if out_path:
+        return normalize_path(out_path)
+    file_path = Path(file_path)
+    anonymized_path = Path(file_path.parent).joinpath(f'{file_path.stem}_anonymized{file_path.suffix}')
+    return normalize_path(anonymized_path)
 
 
 def _get_metadata_path(out_path: str) -> str:
