+++ scripts/anonymize_folder.py
@@ -1,12 +1,24 @@
 import typer
 
-from ips_client.tools.anonymize_folder import anonymize_folder
+from ips_client.data_models import JobArguments, OutputType, ServiceType, Region
+from ips_client.settings import Settings
+from ips_client.tools.anonymize_folder import anonymize_folder as anon_folder, InputTypes
 
 
 app = typer.Typer()
+settings = Settings()
+
+
+@app.command()
+def anonymize_folder(in_dir: str, out_dir: str, input_type: InputTypes, out_type: OutputType, service: ServiceType,
+                     region: Region = Region.european_union, face: bool = True, license_plate: bool = True,
+                     ips_url: str = settings.ips_url_default, n_parallel_jobs: int = 5, save_metadata: bool = True,
+                     skip_existing: bool = True, auto_delete_job: bool = True):
+    job_args = JobArguments(region=region, face=face, license_plate=license_plate)
+    anon_folder(in_dir=in_dir, out_dir=out_dir, input_type=input_type, out_type=out_type, service=service,
+                job_args=job_args, ips_url=ips_url, n_parallel_jobs=n_parallel_jobs, save_metadata=save_metadata,
+                skip_existing=skip_existing, auto_delete_job=auto_delete_job)
 
 
 if __name__ == '__main__':
-    # programmatically decorate function with @app.command
-    app.command()(anonymize_folder)
     app()
